Unalias docker so sourcing file twice doesn't lead to crashes
Since the docker command is used within functions and also aliased by
the script, if the script was sourced twice the wrapper functions would
become broken. This change unaliases docker at the start of the script
to prevent this bug.

We also replace all tabs with spaces and default DOCKER_HOST to
127.0.0.1 in the __dockervnc function.